+++ docs/changes.rst
@@ -7,6 +7,8 @@ Weblate 2026.7
 
 .. rubric:: Improvements
 
+* Management interface access control is now more fine-grained with dedicated site-wide permissions.
+
 .. rubric:: Bug fixes
 
 .. rubric:: Compatibility